requirements.txt
dashscope==1.20.1
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.13
python-dotenv==1.0.1
sparkai==0.3.0
